# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/progress/overview/director": {
      "get": {
        "operationId": "ProgressController_getDirectorOverview",
        "parameters": [
          {
            "name": "period",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeUsers",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Progress"
        ]
      }
    },
    "/api/v1/progress/overview/all": {
      "get": {
        "operationId": "ProgressController_getOverallProgress",
        "parameters": [
          {
            "name": "roles",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Progress"
        ]
      }
    },
    "/api/v1/progress/{userId}": {
      "get": {
        "operationId": "ProgressController_getProgress",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Progress"
        ]
      }
    },
    "/api/v1/progress/bulk": {
      "post": {
        "operationId": "ProgressController_getBulkProgress",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkProgressRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Progress"
        ]
      }
    },
    "/api/v1/progress/assign-roadmap": {
      "post": {
        "operationId": "ProgressController_assignRoadmap",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoadmapDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Progress"
        ]
      }
    },
    "/api/v1/progress/assign-assessment": {
      "post": {
        "operationId": "ProgressController_assignAssessment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignAssessmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Progress"
        ]
      }
    },
    "/api/v1/progress/roadmap/update": {
      "patch": {
        "operationId": "ProgressController_updateRoadmapProgress",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoadmapProgressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Progress"
        ]
      }
    },
    "/api/v1/progress/assessment/update": {
      "patch": {
        "operationId": "ProgressController_updateAssessmentProgress",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssessmentProgressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Progress"
        ]
      }
    },
    "/api/v1/progress/final-comments": {
      "post": {
        "operationId": "ProgressController_addFinalComment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddFinalCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Progress"
        ]
      },
      "patch": {
        "operationId": "ProgressController_updateFinalComment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFinalCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Progress"
        ]
      },
      "delete": {
        "operationId": "ProgressController_deleteFinalComment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteFinalCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Progress"
        ]
      }
    },
    "/api/v1/progress/{userId}/final-comments": {
      "get": {
        "operationId": "ProgressController_getFinalComments",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Progress"
        ]
      }
    },
...
  },
  "info": {
    "title": "CCC Backend API",
    "description": "Christ Church Commission (CCC) Pastor Mentor platform REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT access token"
      }
    },
    "schemas": {
...
      "ObjectId": {
        "type": "object",
        "properties": {}
      },
...
      "BulkProgressRequestDto": {
        "type": "object",
        "properties": {
          "userIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          }
        },
        "required": [
          "userIds"
        ]
      },
      "AssignRoadmapDto": {
        "type": "object",
        "properties": {
          "userIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          "roadMapIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          "assignedBy": {
            "type": "string"
          },
          "dueDate": {
            "type": "string"
          }
        },
        "required": [
          "userIds",
          "roadMapIds"
        ]
      },
      "AssignAssessmentDto": {
        "type": "object",
        "properties": {
          "userIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          "assessmentIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          "dueDate": {
            "type": "string"
          }
        },
        "required": [
          "userIds",
          "assessmentIds"
        ]
      },
      "UpdateRoadmapProgressDto": {
        "type": "object",
        "properties": {
          "userId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "roadMapId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "completedSteps": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "userId",
          "roadMapId",
          "completedSteps"
        ]
      },
      "UpdateAssessmentProgressDto": {
        "type": "object",
        "properties": {
          "userId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "assessmentId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "completedSections": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "userId",
          "assessmentId",
          "completedSections"
        ]
      },
      "AddFinalCommentDto": {
        "type": "object",
        "properties": {
          "userId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "commentorId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "comment": {
            "type": "string",
            "maxLength": 2000
          }
        },
        "required": [
          "userId",
          "commentorId",
          "comment"
        ]
      },
      "UpdateFinalCommentDto": {
        "type": "object",
        "properties": {
          "comment": {
            "type": "string",
            "maxLength": 2000
          },
          "userId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "commentId": {
            "$ref": "#/components/schemas/ObjectId"
          }
        },
        "required": [
          "userId",
          "commentId"
        ]
      },
      "DeleteFinalCommentDto": {
        "type": "object",
        "properties": {
          "userId": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "commentId": {
            "$ref": "#/components/schemas/ObjectId"
          }
        },
        "required": [
          "userId",
          "commentId"
        ]
      },
...
    }
  }
}